+++ tools/linux/git/fetch-fork.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+# Retrieves, Updates and Merge with the fork origin
+git fetch upstream
+git checkout master
+git merge upstream/master