+++ tools/linux/git/pull.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Retrives changes from the origin (not the fork origin, just the repotory's)
+git pull