+++ tools/linux/git/push.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Uploads changes to the origin (not the fork origin, just the repotory's)
+git push