+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
  - Renamed `src/app` to `app/src/module`
 - Reestructure the tools folder
  - Adding support to windows and linux
+- Add git utilities to tools
 
 ## 2022-03-27